+++ tests/ignores.txt
@@ -2,3 +2,4 @@ scrapy/core/downloader/handlers/http.py
 scrapy/extensions/statsmailer.py
 scrapy/interfaces.py
 scrapy/mail.py
+scrapy/utils/console.py
